Undo thor refactoring
This will get brought in through an additional PR.
